% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsel.R
\name{clarity}
\alias{clarity}
\title{Create \link[RSelenium]{remoteDriver} and login to Clarity}
\description{
An R6 Class to manage the Clarity automation
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{clarity$new()}}
\item \href{#method-debug_int}{\code{clarity$debug_int()}}
\item \href{#method-get_hud_export}{\code{clarity$get_hud_export()}}
\item \href{#method-clone}{\code{clarity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clarity$new(
  ini_path = file.path("inst", "auth", "Clarity.ini"),
  clarity_url = "https://cohhio.clarityhs.com/",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clarity_url}}{\code{(character)} the full URL to the Clarity instance to log in to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-debug_int"></a>}}
\if{latex}{\out{\hypertarget{method-debug_int}{}}}
\subsection{Method \code{debug_int()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clarity$debug_int()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_hud_export"></a>}}
\if{latex}{\out{\hypertarget{method-get_hud_export}{}}}
\subsection{Method \code{get_hud_export()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clarity$get_hud_export(
  end_date = Sys.Date(),
  schedule = TRUE,
  browser_dl_folder = "C:\\\\\\\\Users\\\\\\\\Administrator\\\\\\\\Downloads",
  extract_path = "~/R/Contributor_Repos/COHHIO/Rm_data/data",
  delete_archive = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clarity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
