% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsel.R
\docType{data}
\name{dr_env}
\alias{dr_env}
\title{Environment for \link[RSelenium]{remoteDriver}}
\format{
An object of class \code{environment} of length 0.
}
\usage{
dr_env
}
\description{
Environment for \link[RSelenium]{remoteDriver}
}
\keyword{datasets}
