% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsel.R
\name{the_rD}
\alias{the_rD}
\title{Get an instantiated remoteDriver or initiate a new one}
\usage{
the_rD(
  browser = "chrome",
  chromever = "95.0.4638.69",
  port = 5901L,
  e = rlang::caller_env()
)
}
\arguments{
\item{e}{\code{(calling environment)}}
}
\value{
\code{(remoteDriver)}
}
\description{
Get an instantiated remoteDriver or initiate a new one
}
